refactor:  Repository의 ID 타입을 Long으로 변경 및 반환 타입을 Optional로 개선
- 메서드의 반환 타입을 Optional<Password>로 변경하여 결과가 없을 경우를 처리

+++ src/main/java/com/example/kkm/user/auth/repository/UserRepository.java
@@ -1,15 +1,14 @@
 package com.example.kkm.user.auth.repository;
 
-import com.example.kkm.user.auth.entity.User;
+import com.example.kkm.user.domain.entity.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     boolean existsByEmail(String email);
 
     Optional<User> findByEmail(String email);
-
 }
